apps/python/requirements.txt
praw==7.8.1
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.2
httpx==0.28.1
